+++ eoreader/products/optical/maxar_product.py
@@ -1460,12 +1460,6 @@ class MaxarProduct(VhrProduct):
         """
         Compute TOA reflectance from TOA radiance
 
-        See
-        `here <https://apollomapping.com/image_downloads/Maxar_AbsRadCalDataSheet2018v0.pdf>`_
-        for more information.
-
-        WARNING: in this formula, d**2 = 1 / sqrt(dt) !
-
         Args:
             rad_arr (xr.DataArray): TOA Radiance array
             band (BandNames): Band
@@ -1473,11 +1467,7 @@ class MaxarProduct(VhrProduct):
         Returns:
             xr.DataArray: TOA Reflectance array
         """
-
         # Compute the coefficient converting TOA radiance in TOA reflectance
-        dt = self._sun_earth_distance_variation() ** 2
-        _, sun_zen = self.get_mean_sun_angles()
-        rad_sun_zen = np.deg2rad(sun_zen)
         e0 = _MAXAR_E0[self.constellation][band]
 
         if self.constellation == Constellation.WVLG:
@@ -1495,10 +1485,7 @@ class MaxarProduct(VhrProduct):
             else:
                 e0 = e0_lg
 
-        toa_refl_coeff = np.pi / (e0 * dt * np.cos(rad_sun_zen))
-
-        # LOGGER.debug(f"rad to refl coeff = {toa_refl_coeff}")
-        return rad_arr.copy(data=toa_refl_coeff * rad_arr)
+        return self._toa_rad_to_toa_refl_formula(rad_arr, e0)
 
     @cache
     def get_cloud_cover(self) -> float: